restructure tracer eip3155
