+++ bins/revme/src/lib.rs
@@ -1,2 +1,1 @@
 pub mod statetest;
-pub mod tracer_eip3155;
