+++ bins/revme/src/main.rs
@@ -2,7 +2,6 @@ mod cmd;
 mod exec;
 mod runner;
 mod statetest;
-pub mod tracer_eip3155;
 use cmd::Error;
 use structopt::StructOpt;
 mod cli_env;
