+++ crates/revm/Cargo.toml
@@ -17,6 +17,7 @@ auto_impl = { version = "1.0", default-features = false }
 
 # Optional
 serde = { version = "1.0", features = ["derive", "rc"], optional = true }
+serde_json = { version = "1.0", features = ["preserve_order"] }
 
 # ethersdb
 tokio = { version = "1.23", features = [
